+++ src/javascript/constructors/explorer.js
@@ -1,13 +1,25 @@
-import requirePath from '../utils/require'
 import { puffin } from '@mkenzo_8/puffin'
+import requirePath from '../utils/require'
 import Item from '../components/explorer/item'
 import path from 'path'
 import parseDirectory from '../utils/directory.parser'
 import RunningConfig from 'RunningConfig'
+import { saveWorkspace, isWorkspaceLoaded } from '../utils/filesystem'
 
-function Explorer(folderPath,parent,level = 0){
+RunningConfig.on('addFolderToWorkspace',function({
+    path,
+    replaceOldExplorer = true
+}){
+    RunningConfig.data.openedFolders.push({
+        absolutePath:path,
+        folderName:parseDirectory(path)
+    })
+    new Explorer(path,document.getElementById("sidepanel"),0,replaceOldExplorer)
+    if( isWorkspaceLoaded() ) saveWorkspace()
+});
+
+function Explorer(folderPath,parent,level = 0,replaceOldExplorer=true){
     const fs = requirePath('fs-extra')
-
     if(level == 0){
         const explorerContainer = puffin.element(`
             <Item isDirectory="true" parentFolder="${folderPath}" path="${parseDirectory(folderPath)}" fullpath="${folderPath}" level="0"/>
@@ -21,20 +33,24 @@ function Explorer(folderPath,parent,level = 0){
                 }
             }
         })
+        if(replaceOldExplorer && parent.children[0] != null){
 
-        if(parent.children[0] != null){
-            const index = RunningConfig.data.openedFolders.indexOf({
-                absolutePath:parent.children[0].getAttribute("fullpath"),
-                folderName:parseDirectory(parent.children[0].getAttribute("fullpath"))
-            })
-            RunningConfig.data.openedFolders.splice(index,1) //Remove old opened folder from opened folders
+            for( let otherExplorer of parent.children){
+                const index = RunningConfig.data.openedFolders.filter(({absolutePath,folderName},i)=>{
+                    if(absolutePath == otherExplorer.getAttribute("fullpath") && 
+                    folderName == parseDirectory(otherExplorer.getAttribute("fullpath"))) return i
+                })[0]
+
+                RunningConfig.data.openedFolders.splice(index,1) //Remove old opened folder from opened folders
+            }
+           
         }
-
+        
         puffin.render(explorerContainer,parent,{
-            removeContent:true
+            removeContent:replaceOldExplorer
         })
     }
-    
+
     fs.readdir(folderPath).then(function(paths){
         const explorerComponent = puffin.element(`
             <div class="${puffin.style.css`
@@ -66,13 +82,8 @@ function Explorer(folderPath,parent,level = 0){
             puffin.render(explorerComponent,parent,{
                 removeContent:false
             })
-        }else{
-            RunningConfig.data.openedFolders.push({
-                absolutePath:folderPath,
-                folderName:parseDirectory(folderPath)
-            })
         }
-    })    
+    }).catch(err=>console.log(err))
 
     
 }